% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catcat.R
\name{catcat}
\alias{catcat}
\title{CatCat}
\usage{
catcat(A, B)
}
\arguments{
\item{A}{a vector or a CatCat object}

\item{B}{a vector}
}
\value{
a matrix
}
\description{
Create and return a CatCat object with every combination
of \code{A} and \code{B}
}
\examples{
a <- catcat(c(1:3), c(2:8))
b <- catcat(a, c(1:4))
}
