% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catcat.R
\name{catcat_1}
\alias{catcat_1}
\title{CatCat}
\usage{
catcat_1(A, B)
}
\arguments{
\item{A}{a vector}

\item{B}{a vector}
}
\value{
a matrix
}
\description{
Create and return a CatCat object with every combination
of \code{A} and \code{B}
}
\keyword{internal}
