% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hmatrix.R
\name{make_demography_matrix}
\alias{make_demography_matrix}
\title{Make parameters for null human demography model}
\usage{
make_demography_matrix(pars, H)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{H}{size of human population in each strata}
}
\value{
none
}
\description{
Make parameters for null human demography model
}
