% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xds_solve_cohort.R
\name{xds_solve_cohort}
\alias{xds_solve_cohort}
\title{Solve a Dynamical System}
\usage{
xds_solve_cohort(xds_obj, birthday = 0, Amax = 365, da = 1, ages = NULL)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{birthday}{a cohort's birthday}

\item{Amax}{the oldest year, run from 0...Amax}

\item{da}{the age interval}

\item{ages}{a set of ages}
}
\description{
Use \code{xds_setup_eir} to set up a model for cohort dynamics.
This sets the birthday for a cohort \eqn{(d)}, the ages at which output is wanted \eqn{(a)},
and solves for \eqn{t=a+d}:
\deqn{F_w(a) \times F_S(t) \times F_T(t) \times F_K(t)}
}
\note{
The variable \code{xds_obj$EIR_obj$bday} is set up
by \code{xds_setup_eir} but it is set to 0, so age
and time are identical. Using the birthday \eqn{(d)} and
ages \eqn{(A)}, the system calls \code{xds_setup_eir} and
sets \eqn{t = d + A.}
}
